Use sqlx instead of diesel

+++ src/app/v1/users/model/user.rs
@@ -1,9 +1,9 @@
 use chrono::NaiveDateTime;
 use serde::Serialize;
 
-#[derive(Debug, Clone, Eq, PartialEq, Hash, Serialize)]
+#[derive(Debug, Serialize)]
 pub(crate) struct User {
-    pub(crate) id: u32,
+    pub(crate) id: u64,
     pub(crate) name: String,
     pub(crate) email: String,
     pub(crate) created_at: NaiveDateTime,
